Move migration scripts into the common/ directory so we can run them from the launcher.
A few fixes to the launcher paths to get everything writing under data/

+++ build/scripts/migrate_db.rb
@@ -1,17 +1,25 @@
-require 'config/config-distribution'
-
-if (AppConfig[:db_url] =~ /jdbc:derby:(.*?);.*aspacedemo=true$/)
-  dir = $1
-
-  if File.directory?(dir) and File.exists?(File.join(dir, "seg0"))
-    puts "Nuking demo database: #{dir}"
-    sleep(5)
-    FileUtils.rm_rf(dir)
-    exit
-  end
+if $0 =~ /scripts\/rb\/migrate_db\.rb$/
+  # This script runs in two contexts: build/run as a part of development, and
+  # setup-database.(sh|bat) from the distribution zip file.  Allow for both.
+  require_relative '../../launcher/launcher_init'
 end
 
-require_relative "../../backend/app/model/db_migrator"
+require 'config/config-distribution'
+require "db/db_migrator"
+
+
+if ARGV.length > 0 and ARGV[0] == "nuke"
+  if (AppConfig[:db_url] =~ /jdbc:derby:(.*?);.*aspacedemo=true$/)
+    dir = $1
+
+    if File.directory?(dir) and File.exists?(File.join(dir, "seg0"))
+      puts "Nuking demo database: #{dir}"
+      sleep(5)
+      FileUtils.rm_rf(dir)
+      exit
+    end
+  end
+end
 
 
 Sequel.connect(AppConfig[:db_url],
@@ -28,7 +36,7 @@ Sequel.connect(AppConfig[:db_url],
 
   end
 
-  puts "Running migrations..."
+  puts "Running migrations against #{AppConfig[:db_url]}"
   DBMigrator.setup_database(db)
   puts "All done."
 end
